+++ src/emulator/dataconnectEmulator.ts
@@ -3,7 +3,7 @@ import * as childProcess from "child_process";
 import { dataConnectLocalConnString } from "../api";
 import { Constants } from "./constants";
 import { getPID, start, stop, downloadIfNecessary } from "./downloadableEmulators";
-import { EmulatorInfo, EmulatorInstance, Emulators, ListenSpec } from "./types";
+import { EmulatorInfo, EmulatorInstance, Emulators, ListenSpec } from './types';
 import { FirebaseError } from "../error";
 import { EmulatorLogger } from "./emulatorLogger";
 import { RC } from "../rc";
@@ -11,12 +11,13 @@ import { BuildResult, requiresVector } from "../dataconnect/types";
 import { listenSpecsToString } from "./portUtils";
 import { Client, ClientResponse } from "../apiv2";
 import { EmulatorRegistry } from "./registry";
+import { load } from "../dataconnect/load";
 
 export interface DataConnectEmulatorArgs {
-  projectId?: string;
+  projectId: string;
   listen: ListenSpec[];
   configDir: string;
-  locationId?: string;
+  locationId: string;
   auto_download?: boolean;
   rc: RC;
 }
@@ -33,6 +34,8 @@ export interface DataConnectBuildArgs {
 
 export class DataConnectEmulator implements EmulatorInstance {
   private emulatorClient: DataConnectEmulatorClient;
+  // TODO: Smoothly handle cases when the existing emulator is shut down
+  private usingExistingEmulator: boolean = false;  
 
   constructor(private args: DataConnectEmulatorArgs) {
     this.emulatorClient = new DataConnectEmulatorClient();
@@ -60,6 +63,12 @@ export class DataConnectEmulator implements EmulatorInstance {
     } catch (err: any) {
       this.logger.log("DEBUG", `'fdc build' failed with error: ${err.message}`);
     }
+    const alreadyRunning = await this.discoverRunningInstance();
+    if (alreadyRunning) {
+      this.logger.log("INFO", "Detected an instance of the emulator already running with your service, reusing it. This emulator will not be shut down at the end of this command.");
+      this.usingExistingEmulator = true;
+      return;
+    }
     return start(Emulators.DATACONNECT, {
       auto_download: this.args.auto_download,
       listen: listenSpecsToString(this.args.listen),
@@ -73,7 +82,11 @@ export class DataConnectEmulator implements EmulatorInstance {
     return Promise.resolve();
   }
 
-  stop(): Promise<void> {
+  async stop(): Promise<void> {
+    if (this.usingExistingEmulator) {
+      this.logger.log("INFO", "Skipping cleanup of Data Connect emulator, as it was not started by this process.");
+      return;
+    }
     return stop(Emulators.DATACONNECT);
   }
 
@@ -144,6 +157,23 @@ export class DataConnectEmulator implements EmulatorInstance {
     return this.args.rc.getDataconnect()?.postgres?.localConnectionString;
   }
 
+  private async discoverRunningInstance(): Promise<boolean> {
+     const emuInfo = await this.emulatorClient.getInfo();
+     if (!emuInfo) {
+      return false;
+     }
+     const serviceInfo = await load(this.args.projectId, this.args.locationId,this.args.configDir);
+     const sameService =  emuInfo.services.some(s => serviceInfo.dataConnectYaml.serviceId === s.serviceId);
+     if (!sameService) {
+      throw new FirebaseError(`There is a Data Connect emulator already running on ${
+        this.args.listen[0].address
+      }:${
+        this.args.listen[0].port
+      }, but it is emulating a different service. Please stop that instance of the Data Connect emulator, or specify a different port in 'firebase.json'`);
+     }
+     return true;
+  }
+
   public async connectToPostgres(
     localConnectionString?: string,
     database?: string,
@@ -171,6 +201,20 @@ type ConfigureEmulatorRequest = {
   database?: string;
 };
 
+
+type GetInfoResponse = {
+  // Version number of the emulator.
+  version: string,
+  // List of services currently running on the emulator.
+  services: {
+    // ID of this service.
+    serviceId: string,
+    // The Postgres connection string that this service uses.
+    connectionString: string,
+  }[]
+};
+
+
 export class DataConnectEmulatorClient {
   private client: Client | undefined = undefined;
 
@@ -190,5 +234,31 @@ export class DataConnectEmulatorClient {
       }
       throw err;
     }
+  }
+
+  public async getInfo(): Promise<GetInfoResponse | void> {
+    if (!this.client) {
+      this.client = EmulatorRegistry.client(Emulators.DATACONNECT);
+    }
+    return getInfo(this.client);
+  }
+}
+
+export async function checkIfDataConnectEmulatorRunningOnAddress(l: ListenSpec) {
+  const client = new Client({
+    urlPrefix: `http:/${l.family === "IPv6" ? `[${l.address}]` : l.address }:${l.port}`,
+    auth: false,
+  })
+  return getInfo(client);
+}
+
+async function getInfo(client: Client): Promise<GetInfoResponse | void> {
+  try {
+    const res = await client.get<GetInfoResponse>(
+      "emulator/info",
+    );
+    return res.body;
+  } catch (err) {
+    return;
   }
 }